% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{encode}
\alias{encode}
\title{Encode the Source File for Calculating}
\usage{
encode(path, user_visit_lower_bound = 0L, progress = 0, sep1 = "\\001",
  sep2 = "\\002", sep3 = "\\003")
}
\arguments{
\item{path}{string. The path to the source file.}

\item{user_visit_lower_bound}{integer. If the user visit is less than
\code{user_visit_lower_bound}, then he/she will be filtered out.}

\item{progress}{numeric. The number of lines of the source file. If
the progress is set non-zero, then a progress bar will be displayed.}

\item{sep1}{single character. The delimiter between the user and the rest.}

\item{sep2}{single character. The delimiter between the pairs of item and the count.}

\item{sep3}{single character. The delimiter between the item and count.}
}
\description{
Encode the Source File for Calculating
}

