% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/train.R
\name{train_pmf}
\alias{train_pmf}
\title{Train a Poisson Matrix Factorization Model}
\usage{
train_pmf(src, prior, output)
}
\arguments{
\item{src}{path. Please see details for more information.}

\item{prior}{list. A named numeric vector of prior.}

\item{output}{path to a directory. If the path does not exist, R will try
to create the directory. The meta data and model will be written to the
directory.}
}
\description{
Train a Poisson Matrix Factorization Model
}
\details{
The Poisson Matrix Factorization(PMF) model assumes that the counting response
\eqn{y_{u,i}} corresponding to the user $u$ and item $i$ is poisson distributed
with \eqn{Ey_{u,i} = \sum_{k=1}^K {\theta_{u,k} \beta_{i,k}}}.
}

